more docs and code-rearrange
